+++ app/controllers/favorites_controller.rb
@@ -83,8 +83,11 @@ class FavoritesController < ApplicationController
     the_id = params.fetch("path_id")
     the_favorite = Favorite.where({ :id => the_id }).at(0)
 
-    the_favorite.destroy
-
-    redirect_to("/favorites", { :notice => "Favorite deleted successfully."} )
+    if @current_user && the_favorite.user_id == @current_user.id
+      the_favorite.destroy
+      redirect_to("/favorites", { :notice => "Favorite deleted successfully."} )
+    else
+      redirect_to("/user_sign_in", { :alert => "You must be logged in to perform this action."})
+    end
   end
 end
